Updated to 0.0.9
