+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@eth-optimism/plugins",
-  "version": "0.0.9-alpha.2",
+  "version": "0.0.9",
   "main": "index.js",
   "types": "index.d.ts",
   "files": [
@@ -28,6 +28,7 @@
     "@types/chai": "^4.2.14",
     "@types/mocha": "^8.2.0",
     "@types/node": "^14.14.19",
+    "@types/node-fetch": "^2.5.8",
     "@types/sinon-chai": "^3.2.5",
     "chai": "^4.2.0",
     "ethereum-waffle": "^3.0.0",
@@ -48,7 +49,8 @@
     "bn.js": "^5.1.3",
     "ethereumjs-account": "^3.0.0",
     "ethjs-common-v1": "npm:ethereumjs-common@1.5.0",
-    "ethjs-util-v6": "npm:ethereumjs-util@6.2.1"
+    "ethjs-util-v6": "npm:ethereumjs-util@6.2.1",
+    "node-fetch": "^2.6.1"
   },
   "peerDependencies": {
     "@nomiclabs/hardhat-ethers": "^2.0.1",
