--- src/hardhat/compiler/0.5.16/index.ts
@@ -1,3 +1,0 @@
-import { loadCompilerPlugin } from '../internal'
-
-loadCompilerPlugin('0.5.16')
