--- src/hardhat/compiler/0.6.12/index.ts
@@ -1,3 +1,0 @@
-import { loadCompilerPlugin } from '../internal'
-
-loadCompilerPlugin('0.6.12')
