--- src/hardhat/compiler/0.7.6/index.ts
@@ -1,3 +1,0 @@
-import { loadCompilerPlugin } from '../internal'
-
-loadCompilerPlugin('0.7.6')
