+++ src/hardhat/compiler/type-extensions.ts
@@ -1,0 +1,15 @@
+import 'hardhat/types/config'
+
+declare module 'hardhat/types/config' {
+  interface HardhatUserConfig {
+    ovm?: {
+      solcVersion?: string
+    }
+  }
+
+  interface HardhatConfig {
+    ovm?: {
+      solcVersion?: string
+    }
+  }
+}
